refactor: member 패키지 port/adapter 구조로 변경
