+++ src/main/java/dandi/dandi/auth/application/AuthService.java
@@ -1,46 +1,49 @@
 package dandi.dandi.auth.application;
 
-import dandi.dandi.auth.application.dto.LoginRequest;
-import dandi.dandi.auth.application.dto.LoginResponse;
-import dandi.dandi.auth.application.dto.TokenResponse;
+import dandi.dandi.auth.application.port.in.AuthUseCase;
+import dandi.dandi.auth.application.port.in.dto.LoginRequest;
+import dandi.dandi.auth.application.port.out.RefreshTokenPersistencePort;
+import dandi.dandi.auth.application.port.out.dto.LoginResponse;
+import dandi.dandi.auth.application.port.out.dto.TokenResponse;
 import dandi.dandi.auth.domain.OAuthClient;
 import dandi.dandi.auth.domain.RefreshToken;
-import dandi.dandi.auth.domain.RefreshTokenRepository;
 import dandi.dandi.auth.exception.UnauthorizedException;
 import dandi.dandi.auth.infrastructure.token.JwtTokenManager;
 import dandi.dandi.auth.infrastructure.token.RefreshTokenManager;
+import dandi.dandi.member.application.port.out.MemberPersistencePort;
+import dandi.dandi.member.application.service.MemberSignupManager;
 import dandi.dandi.member.domain.Member;
-import dandi.dandi.member.domain.MemberRepository;
-import dandi.dandi.member.domain.MemberSignupManager;
 import java.util.Optional;
 import javax.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
 @Service
-public class AuthService {
+public class AuthService implements AuthUseCase {
 
     private final MemberSignupManager memberSignupManager;
     private final OAuthClient oAuthClient;
-    private final MemberRepository memberRepository;
-    private final RefreshTokenRepository refreshTokenRepository;
+    private final MemberPersistencePort memberPersistencePort;
+    private final RefreshTokenPersistencePort refreshTokenPersistencePort;
     private final RefreshTokenManager refreshTokenManager;
     private final JwtTokenManager jwtTokenManager;
 
     public AuthService(MemberSignupManager memberSignupManager, OAuthClient oAuthClient,
-                       MemberRepository memberRepository, RefreshTokenRepository refreshTokenRepository,
+                       MemberPersistencePort memberPersistencePort,
+                       RefreshTokenPersistencePort refreshTokenPersistencePort,
                        RefreshTokenManager refreshTokenManager, JwtTokenManager jwtTokenManager) {
         this.memberSignupManager = memberSignupManager;
         this.oAuthClient = oAuthClient;
-        this.memberRepository = memberRepository;
-        this.refreshTokenRepository = refreshTokenRepository;
+        this.memberPersistencePort = memberPersistencePort;
+        this.refreshTokenPersistencePort = refreshTokenPersistencePort;
         this.refreshTokenManager = refreshTokenManager;
         this.jwtTokenManager = jwtTokenManager;
     }
 
+    @Override
     @Transactional
     public LoginResponse getToken(LoginRequest loginRequest) {
         String oAuthMemberId = oAuthClient.getOAuthMemberId(loginRequest.getIdToken());
-        Optional<Member> member = memberRepository.findByOAuthId(oAuthMemberId);
+        Optional<Member> member = memberPersistencePort.findByOAuthId(oAuthMemberId);
         if (member.isPresent()) {
             String accessToken = jwtTokenManager.generateToken(String.valueOf(member.get().getId()));
             String refreshToken = generateRefreshToken(member.get().getId());
@@ -58,13 +61,14 @@ public class AuthService {
 
     private String generateRefreshToken(Long memberId) {
         RefreshToken refreshToken = refreshTokenManager.generateToken(memberId);
-        refreshTokenRepository.save(refreshToken);
+        refreshTokenPersistencePort.save(refreshToken);
         return refreshToken.getValue();
     }
 
+    @Override
     @Transactional
     public TokenResponse refresh(Long memberId, String refreshToken) {
-        RefreshToken refreshTokenInfo = refreshTokenRepository
+        RefreshToken refreshTokenInfo = refreshTokenPersistencePort
                 .findRefreshTokenByMemberIdAndValue(memberId, refreshToken)
                 .orElseThrow(UnauthorizedException::refreshTokenNotFound);
         validateExpiration(refreshTokenInfo);
@@ -79,8 +83,9 @@ public class AuthService {
         }
     }
 
+    @Override
     @Transactional
     public void logout(Long memberId) {
-        refreshTokenRepository.deleteByMemberId(memberId);
+        refreshTokenPersistencePort.deleteByMemberId(memberId);
     }
 }
