--- src/main/java/dandi/dandi/auth/application/dto/LoginRequest.java
@@ -1,20 +1,0 @@
-package dandi.dandi.auth.application.dto;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-
-public class LoginRequest {
-
-    @Schema(example = "Identity Token From Apple Server")
-    private String idToken;
-
-    public LoginRequest() {
-    }
-
-    public LoginRequest(String idToken) {
-        this.idToken = idToken;
-    }
-
-    public String getIdToken() {
-        return idToken;
-    }
-}
