--- src/main/java/dandi/dandi/auth/domain/RefreshTokenRepository.java
@@ -1,11 +1,0 @@
-package dandi.dandi.auth.domain;
-
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
-
-    Optional<RefreshToken> findRefreshTokenByMemberIdAndValue(Long memberId, String value);
-
-    void deleteByMemberId(Long memberId);
-}
