--- src/main/java/dandi/dandi/member/application/dto/NicknameUpdateRequest.java
@@ -1,20 +1,0 @@
-package dandi.dandi.member.application.dto;
-
-import io.swagger.v3.oas.annotations.media.Schema;
-
-public class NicknameUpdateRequest {
-
-    @Schema(example = "new Nickname")
-    private String newNickname;
-
-    public NicknameUpdateRequest() {
-    }
-
-    public NicknameUpdateRequest(String newNickname) {
-        this.newNickname = newNickname;
-    }
-
-    public String getNewNickname() {
-        return newNickname;
-    }
-}
