--- src/main/java/dandi/dandi/member/application/dto/ProfileImageUpdateResponse.java
@@ -1,17 +1,0 @@
-package dandi.dandi.member.application.dto;
-
-public class ProfileImageUpdateResponse {
-
-    private String profileImageUrl;
-
-    public ProfileImageUpdateResponse() {
-    }
-
-    public ProfileImageUpdateResponse(String profileImageUrl) {
-        this.profileImageUrl = profileImageUrl;
-    }
-
-    public String getProfileImageUrl() {
-        return profileImageUrl;
-    }
-}
