+++ src/main/java/dandi/dandi/member/domain/NewMemberCreatedEvent.java
@@ -1,5 +1,7 @@
 package dandi.dandi.member.domain;
 
+import java.util.Objects;
+
 public class NewMemberCreatedEvent {
 
     private final Long memberId;
@@ -10,5 +12,22 @@ public class NewMemberCreatedEvent {
 
     public Long getMemberId() {
         return memberId;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof NewMemberCreatedEvent)) {
+            return false;
+        }
+        NewMemberCreatedEvent that = (NewMemberCreatedEvent) o;
+        return Objects.equals(memberId, that.memberId);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(memberId);
     }
 }
