+++ src/test/java/dandi/dandi/member/application/MemberServiceTest.java
@@ -3,33 +3,33 @@ package dandi.dandi.member.application;
 import static dandi.dandi.member.MemberTestFixture.INITIAL_PROFILE_IMAGE_URL;
 import static dandi.dandi.member.MemberTestFixture.NICKNAME;
 import static dandi.dandi.member.MemberTestFixture.OAUTH_ID;
+import static dandi.dandi.member.MemberTestFixture.TEST_MEMBER;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
-import static org.mockito.Mockito.only;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import dandi.dandi.auth.exception.UnauthorizedException;
-import dandi.dandi.member.application.dto.LocationUpdateRequest;
-import dandi.dandi.member.application.dto.MemberInfoResponse;
-import dandi.dandi.member.application.dto.NicknameUpdateRequest;
+import dandi.dandi.member.application.port.in.dto.LocationUpdateRequest;
+import dandi.dandi.member.application.port.in.dto.NicknameUpdateRequest;
+import dandi.dandi.member.application.port.out.MemberPersistencePort;
+import dandi.dandi.member.application.port.out.dto.MemberInfoResponse;
+import dandi.dandi.member.application.service.MemberService;
 import dandi.dandi.member.domain.Member;
-import dandi.dandi.member.domain.MemberRepository;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
 class MemberServiceTest {
 
     @Mock
-    private MemberRepository memberRepository;
+    private MemberPersistencePort memberPersistencePort;
 
     @InjectMocks
     private MemberService memberService;
@@ -38,7 +38,7 @@ class MemberServiceTest {
     @Test
     void findMemberInfo() {
         Long memberId = 1L;
-        when(memberRepository.findById(memberId))
+        when(memberPersistencePort.findById(memberId))
                 .thenReturn(Optional.of(Member.initial(OAUTH_ID, NICKNAME, INITIAL_PROFILE_IMAGE_URL)));
 
         MemberInfoResponse memberInfoResponse = memberService.findMemberInfo(memberId);
@@ -55,7 +55,7 @@ class MemberServiceTest {
     @Test
     void findMemberInfo_NonExistentMember() {
         Long nonExistentMemberId = 1L;
-        when(memberRepository.findById(nonExistentMemberId))
+        when(memberPersistencePort.findById(nonExistentMemberId))
                 .thenReturn(Optional.empty());
 
         assertThatThrownBy(() -> memberService.findMemberInfo(nonExistentMemberId))
@@ -66,31 +66,27 @@ class MemberServiceTest {
     @DisplayName("회원의 닉네임을 변경할 수 있다.")
     @Test
     void updateNickname() {
-        Long memberId = 1L;
-        Member member = Mockito.mock(Member.class);
         String newNickname = "newNickname";
         NicknameUpdateRequest nicknameUpdateRequest = new NicknameUpdateRequest(newNickname);
-        when(memberRepository.findById(memberId))
-                .thenReturn(Optional.of(member));
+        when(memberPersistencePort.findById(TEST_MEMBER.getId()))
+                .thenReturn(Optional.of(TEST_MEMBER));
 
-        memberService.updateNickname(memberId, nicknameUpdateRequest);
+        memberService.updateNickname(TEST_MEMBER.getId(), nicknameUpdateRequest);
 
-        verify(member, only()).updateNickname(newNickname);
+        verify(memberPersistencePort).updateNickname(TEST_MEMBER.getId(), newNickname);
     }
 
     @DisplayName("회원의 위치 정보를 변경할 수 있다.")
     @Test
     void updateLocation() {
-        Long memberId = 1L;
-        Member member = Mockito.mock(Member.class);
         double latitude = 1.0;
         double longitude = 1.0;
         LocationUpdateRequest locationUpdateRequest = new LocationUpdateRequest(latitude, longitude);
-        when(memberRepository.findById(memberId))
-                .thenReturn(Optional.of(member));
+        when(memberPersistencePort.findById(TEST_MEMBER.getId()))
+                .thenReturn(Optional.of(TEST_MEMBER));
 
-        memberService.updateLocation(memberId, locationUpdateRequest);
+        memberService.updateLocation(TEST_MEMBER.getId(), locationUpdateRequest);
 
-        verify(member, only()).updateLocation(latitude, latitude);
+        verify(memberPersistencePort).updateLocation(TEST_MEMBER.getId(), latitude, latitude);
     }
 }
