+++ src/test/java/dandi/dandi/member/application/ProfileImageServiceTest.java
@@ -18,9 +18,10 @@ import static org.mockito.Mockito.when;
 import com.amazonaws.SdkClientException;
 import dandi.dandi.auth.exception.UnauthorizedException;
 import dandi.dandi.image.exception.ImageUploadFailedException;
+import dandi.dandi.member.application.port.out.MemberPersistencePort;
+import dandi.dandi.member.application.port.out.ProfileImageUploader;
+import dandi.dandi.member.application.service.ProfileImageService;
 import dandi.dandi.member.domain.Member;
-import dandi.dandi.member.domain.MemberRepository;
-import dandi.dandi.member.domain.ProfileImageUploader;
 import java.io.IOException;
 import java.util.Optional;
 import java.util.stream.Stream;
@@ -40,9 +41,9 @@ class ProfileImageServiceTest {
 
     private static final String TEST_PROFILE_BUCKET_IMG_DIR = "test-dir";
 
-    private final MemberRepository memberRepository = Mockito.mock(MemberRepository.class);
+    private final MemberPersistencePort memberPersistencePort = Mockito.mock(MemberPersistencePort.class);
     private final ProfileImageUploader profileImageUploader = Mockito.mock(ProfileImageUploader.class);
-    private final ProfileImageService profileImageService = new ProfileImageService(memberRepository,
+    private final ProfileImageService profileImageService = new ProfileImageService(memberPersistencePort,
             profileImageUploader, INITIAL_PROFILE_IMAGE_URL, TEST_PROFILE_BUCKET_IMG_DIR);
 
     @DisplayName("기본 프로필 사진이 아닌 회원의 프로필 사진을 변경하면 기존 사진을 삭제하고 사진을 업로드 한 후, 사진의 식별값을 반환한다.")
@@ -51,7 +52,7 @@ class ProfileImageServiceTest {
         Long memberId = 1L;
         Member notInitialProfileImageMember =
                 Member.initial(OAUTH_ID, NICKNAME, NOT_INITIAL_PROFILE_IMAGE_URL);
-        when(memberRepository.findById(memberId))
+        when(memberPersistencePort.findById(memberId))
                 .thenReturn(Optional.of(notInitialProfileImageMember));
 
         String imgUrl = profileImageService.updateProfileImage(memberId, generateTestImgMultipartFile())
@@ -67,7 +68,7 @@ class ProfileImageServiceTest {
     @Test
     void updateProfileImage_InitialProfileImageMember() throws IOException {
         Long memberId = 1L;
-        when(memberRepository.findById(memberId))
+        when(memberPersistencePort.findById(memberId))
                 .thenReturn(Optional.of(Member.initial(OAUTH_ID, NICKNAME, INITIAL_PROFILE_IMAGE_URL)));
 
         String imgUrl = profileImageService.updateProfileImage(memberId, generateTestImgMultipartFile())
@@ -84,7 +85,7 @@ class ProfileImageServiceTest {
     @MethodSource("provideImageUploadAvailableException")
     void updateProfileImageUrl_ImageUploadFailed(Exception exception) throws IOException {
         Long memberId = 1L;
-        when(memberRepository.findById(memberId))
+        when(memberPersistencePort.findById(memberId))
                 .thenReturn(Optional.of(Member.initial(OAUTH_ID, NICKNAME, INITIAL_PROFILE_IMAGE_URL)));
         doThrow(exception)
                 .when(profileImageUploader)
@@ -105,7 +106,7 @@ class ProfileImageServiceTest {
     @Test
     void updateProfileImage_notExistentMember() {
         Long notExistentMemberId = 1L;
-        when(memberRepository.findById(notExistentMemberId))
+        when(memberPersistencePort.findById(notExistentMemberId))
                 .thenReturn(Optional.empty());
 
         assertThatThrownBy(
